+++ src/OpenTelemetry.Exporter.Geneva/GenevaLogExporter.cs
@@ -31,6 +31,7 @@ namespace OpenTelemetry.Exporter.Geneva;
 public class GenevaLogExporter : GenevaBaseExporter<LogRecord>
 {
     private const int BUFFER_SIZE = 65360; // the maximum ETW payload (inclusive)
+    private const int MaxSanitizedEventNameLength = 50;
 
     private readonly IReadOnlyDictionary<string, object> m_customFields;
     private readonly string m_defaultEventName = "Log";
@@ -44,6 +45,7 @@ public class GenevaLogExporter : GenevaBaseExporter<LogRecord>
     };
 
     private readonly IDataTransport m_dataTransport;
+    private readonly bool shouldPassThruTableMappings;
     private bool isDisposed;
     private Func<object, string> convertToJson;
 
@@ -65,7 +67,14 @@ public class GenevaLogExporter : GenevaBaseExporter<LogRecord>
 
                 if (kv.Key == "*")
                 {
-                    this.m_defaultEventName = kv.Value;
+                    if (kv.Value == "*")
+                    {
+                        this.shouldPassThruTableMappings = true;
+                    }
+                    else
+                    {
+                        this.m_defaultEventName = kv.Value;
+                    }
                 }
                 else
                 {
@@ -204,14 +213,6 @@ public class GenevaLogExporter : GenevaBaseExporter<LogRecord>
             listKvp = logRecord.State as IReadOnlyList<KeyValuePair<string, object>>;
         }
 
-        var name = logRecord.CategoryName;
-
-        // If user configured explicit TableName, use it.
-        if (this.m_tableMappings == null || !this.m_tableMappings.TryGetValue(name, out var eventName))
-        {
-            eventName = this.m_defaultEventName;
-        }
-
         var buffer = m_buffer.Value;
         if (buffer == null)
         {
@@ -242,7 +243,43 @@ public class GenevaLogExporter : GenevaBaseExporter<LogRecord>
         var timestamp = logRecord.Timestamp;
         var cursor = 0;
         cursor = MessagePackSerializer.WriteArrayHeader(buffer, cursor, 3);
-        cursor = MessagePackSerializer.SerializeAsciiString(buffer, cursor, eventName);
+
+        var categoryName = logRecord.CategoryName;
+        string eventName = null;
+
+        Span<byte> sanitizedEventName = default;
+
+        // If user configured explicit TableName, use it.
+        if (this.m_tableMappings != null && this.m_tableMappings.TryGetValue(categoryName, out eventName))
+        {
+            cursor = MessagePackSerializer.SerializeAsciiString(buffer, cursor, eventName);
+        }
+        else if (!this.shouldPassThruTableMappings)
+        {
+            eventName = this.m_defaultEventName;
+            cursor = MessagePackSerializer.SerializeAsciiString(buffer, cursor, eventName);
+        }
+        else
+        {
+            int cursorStartIdx = cursor;
+
+            if (categoryName.Length > 0)
+            {
+                cursor = SerializeSanitizedCategoryName(buffer, cursor, categoryName);
+            }
+
+            if (cursor == cursorStartIdx)
+            {
+                // Serializing null as categoryName could not be sanitized into a valid string.
+                cursor = MessagePackSerializer.SerializeNull(buffer, cursor);
+            }
+            else
+            {
+                // Sanitized category name has been serialized.
+                sanitizedEventName = buffer.AsSpan().Slice(cursorStartIdx, cursor - cursorStartIdx);
+            }
+        }
+
         cursor = MessagePackSerializer.WriteArrayHeader(buffer, cursor, 1);
         cursor = MessagePackSerializer.WriteArrayHeader(buffer, cursor, 2);
         cursor = MessagePackSerializer.SerializeUtcDateTime(buffer, cursor, timestamp);
@@ -282,7 +319,15 @@ public class GenevaLogExporter : GenevaBaseExporter<LogRecord>
         }
 
         // Part A - core envelope
-        cursor = AddPartAField(buffer, cursor, Schema.V40.PartA.Name, eventName);
+        if (sanitizedEventName.Length != 0)
+        {
+            cursor = AddPartAField(buffer, cursor, Schema.V40.PartA.Name, sanitizedEventName);
+        }
+        else
+        {
+            cursor = AddPartAField(buffer, cursor, Schema.V40.PartA.Name, eventName);
+        }
+
         cntFields += 1;
 
         cursor = AddPartAField(buffer, cursor, Schema.V40.PartA.Time, timestamp);
@@ -329,7 +374,7 @@ public class GenevaLogExporter : GenevaBaseExporter<LogRecord>
         cntFields += 1;
 
         cursor = MessagePackSerializer.SerializeAsciiString(buffer, cursor, "name");
-        cursor = MessagePackSerializer.SerializeUnicodeString(buffer, cursor, name);
+        cursor = MessagePackSerializer.SerializeUnicodeString(buffer, cursor, categoryName);
         cntFields += 1;
 
         bool hasEnvProperties = false;
@@ -442,6 +487,61 @@ public class GenevaLogExporter : GenevaBaseExporter<LogRecord>
             default:
                 return 1;
         }
+    }
+
+    // This method would map the logger category to a table name which only contains alphanumeric values with the following additions:
+    // Any character that is not allowed will be removed.
+    // If the resulting string is longer than 50 characters, only the first 50 characters will be taken.
+    // If the first character in the resulting string is a lower-case alphabet, it will be converted to the corresponding upper-case.
+    // If the resulting string still does not comply with Rule, the category name will not be serialized.
+    private static int SerializeSanitizedCategoryName(byte[] buffer, int cursor, string categoryName)
+    {
+        int cursorStartIdx = cursor;
+
+        // Reserve 2 bytes for storing LIMIT_MAX_STR8_LENGTH_IN_BYTES and (byte)validNameLength -
+        // these 2 bytes will be back filled after iterating through categoryName.
+        cursor += 2;
+        int validNameLength = 0;
+
+        // Special treatment for the first character.
+        var firstChar = categoryName[0];
+        if (firstChar >= 'A' && firstChar <= 'Z')
+        {
+            buffer[cursor++] = (byte)firstChar;
+            ++validNameLength;
+        }
+        else if (firstChar >= 'a' && firstChar <= 'z')
+        {
+            // If the first character in the resulting string is a lower-case alphabet,
+            // it will be converted to the corresponding upper-case.
+            buffer[cursor++] = (byte)(firstChar - 32);
+            ++validNameLength;
+        }
+        else
+        {
+            // Not a valid name.
+            return cursor -= 2;
+        }
+
+        for (int i = 1; i < categoryName.Length; ++i)
+        {
+            if (validNameLength == MaxSanitizedEventNameLength)
+            {
+                break;
+            }
+
+            var cur = categoryName[i];
+            if ((cur >= 'a' && cur <= 'z') || (cur >= 'A' && cur <= 'Z') || (cur >= '0' && cur <= '9'))
+            {
+                buffer[cursor++] = (byte)cur;
+                ++validNameLength;
+            }
+        }
+
+        // Backfilling MessagePack serialization protocol and valid category length to the startIdx of the categoryName byte array.
+        MessagePackSerializer.WriteStr8Header(buffer, cursorStartIdx, validNameLength);
+
+        return cursor;
     }
 }
 #endif