+++ src/OpenTelemetry.Extensions.PersistentStorage/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog - OpenTelemetry.Extensions.PersistentStorage
 
+## Unreleased
+
+* [Update implementation to use
+  Opentelemetry.Extensions.PersistentStorage.Abstractions](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/363)
+
 ## 1.0.0-alpha.3
 
 * Going forward the NuGet package will be