+++ src/OpenTelemetry.Instrumentation.ElasticsearchClient/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+* Updated OTel SDK package version to 1.2.0
+* Update minimum full framework support to net462
+* Requests that get an HTTP status code of 404 are not marked as an error span status
+* Add MaxDbStatementLength option with default of 4096
+* Remove duplicated HTTP method and URL from db.statement attribute value
+
 ## 1.0.0-beta.3
 
 * Going forward the NuGet package will be